+++ code/src/nuvla/job/actions/nuvlabox_check_api.py
@@ -6,7 +6,7 @@ from ..actions import action
 from nuvla.connector import nuvlabox_connector as NB
 
 
-@action('check_nuvlabox_api')
+@action('check_nuvlabox_api', True)
 class NBCheckJob(object):
 
     def __init__(self, _, job):