+++ code/tests/job_test.py
@@ -2,6 +2,7 @@ import unittest
 from mock import Mock
 
 from nuvla.job import job
+from nuvla.api.models import CimiResource
 
 
 class TestJob(unittest.TestCase):
@@ -14,7 +15,7 @@ class TestJob(unittest.TestCase):
         # The job will be processed until engine's version is below 2.0.0.
         for ev in ['0.0.1', '1.0.0', '1.2.3']:
             job.engine_version = ev
-            job.Job.get_cimi_job = Mock(return_value={})
+            job.Job.get_cimi_job = Mock(return_value=CimiResource({}))
             job.Job.update_job = Mock()
             jb = job.Job(None, queue)
             assert jb.nothing_to_do is False
@@ -22,7 +23,7 @@ class TestJob(unittest.TestCase):
         # No job version provided with the job. It will be set to the min value.
         # The job will not be processed and removed from the queue.
         job.engine_version = '2.0.0'
-        job.Job.get_cimi_job = Mock(return_value={})
+        job.Job.get_cimi_job = Mock(return_value=CimiResource({}))
         job.Job.update_job = Mock()
         job.retry_kazoo_queue_op = Mock()
         jb = job.Job(None, queue)
@@ -33,7 +34,7 @@ class TestJob(unittest.TestCase):
         # Job version is strictly smaller than M-1 of job engine version.
         # The job will not be processed and removed from the queue.
         job.engine_version = '3.2.1'
-        job.Job.get_cimi_job = Mock(return_value={'version': '1.2.3'})
+        job.Job.get_cimi_job = Mock(return_value=CimiResource({'version': '1.2.3'}))
         job.Job.update_job = Mock()
         job.retry_kazoo_queue_op = Mock()
         jb = job.Job(None, queue)
@@ -44,7 +45,7 @@ class TestJob(unittest.TestCase):
         # The job version might only indicate the major number.
         # Job version is strictly smaller than M-1 of job engine version.
         job.engine_version = '3.2.1'
-        job.Job.get_cimi_job = Mock(return_value={'version': '1'})
+        job.Job.get_cimi_job = Mock(return_value=CimiResource({'version': '1'}))
         job.Job.update_job = Mock()
         job.retry_kazoo_queue_op = Mock()
         jb = job.Job(None, queue)
@@ -55,7 +56,7 @@ class TestJob(unittest.TestCase):
         # The job version might only indicate the major number.
         # Versions are equal. Job will be processed.
         job.engine_version = '1.0.0'
-        job.Job.get_cimi_job = Mock(return_value={'version': '1'})
+        job.Job.get_cimi_job = Mock(return_value=CimiResource({'version': '1'}))
         job.Job.update_job = Mock()
         jb = job.Job(None, queue)
         assert jb.nothing_to_do is False
@@ -64,7 +65,7 @@ class TestJob(unittest.TestCase):
         # Job will be processed.
         for v in ['0.0.1', '1.0.0', '7.6.5']:
             job.engine_version = v
-            job.Job.get_cimi_job = Mock(return_value={'version': v})
+            job.Job.get_cimi_job = Mock(return_value=CimiResource({'version': v}))
             job.Job.update_job = Mock()
             job.retry_kazoo_queue_op = Mock()
             jb = job.Job(None, queue)
@@ -75,21 +76,21 @@ class TestJob(unittest.TestCase):
         # Job's version is higher that engine's.
         # Job is put back to the queue.
         job.engine_version = '0.0.1'
-        job.Job.get_cimi_job = Mock(return_value={'version': '0.0.2'})
+        job.Job.get_cimi_job = Mock(return_value=CimiResource({'version': '0.0.2'}))
         job.retry_kazoo_queue_op = Mock()
         jb = job.Job(None, queue)
         job.retry_kazoo_queue_op.assert_called_once_with(queue, "release")
         assert jb.nothing_to_do is True
 
         job.engine_version = '0.0.1-SNAPSHOT'
-        job.Job.get_cimi_job = Mock(return_value={'version': '0.0.2'})
+        job.Job.get_cimi_job = Mock(return_value=CimiResource({'version': '0.0.2'}))
         job.retry_kazoo_queue_op = Mock()
         jb = job.Job(None, queue)
         job.retry_kazoo_queue_op.assert_called_once_with(queue, "release")
         assert jb.nothing_to_do is True
 
         job.engine_version = '7.6.5'
-        job.Job.get_cimi_job = Mock(return_value={'version': '8.0.1'})
+        job.Job.get_cimi_job = Mock(return_value=CimiResource({'version': '8.0.1'}))
         job.retry_kazoo_queue_op = Mock()
         jb = job.Job(None, queue)
         job.retry_kazoo_queue_op.assert_called_once_with(queue, "release")