+++ container/container-build.sh
@@ -27,6 +27,7 @@ for platform in "${platforms[@]}"; do
            --opt build-arg:GIT_DIRTY=${GIT_DIRTY} \
            --opt build-arg:TRAVIS_BUILD_NUMBER=${TRAVIS_BUILD_NUMBER} \
            --opt build-arg:TRAVIS_BUILD_WEB_URL=${TRAVIS_BUILD_WEB_URL} \
+           --opt build-arg:IMAGE_NAME=${DOCKER_IMAGE} \
            --output type=docker,name=${MANIFEST}-${platform},dest=/tmp/work/target/${DOCKER_IMAGE}-${platform}.docker.tar \
            --local context=/tmp/work \
            --local dockerfile=/tmp/work \