+++ CHANGELOG.md
@@ -6,6 +6,13 @@
 
  - New distributor for setting nuvlabox status to offline
  - New Job Engine Lite Docker image build, optimized for running jobs from within an infrastructure, with multi-arch support
+ - When a job with execution-mode=mixed fails, it get back in the job queue in pull-mode
+ - Create deployment_state jobs with execution-mode 
+ - Add pause entrypoint
+ 
+### Changed
+
+ - Credential check - support the UNKNOWN status for a credential check
  
 ## [2.10.0] - 2020-12-10